atlys: Cleanup and modernize.
