+++ targets/atlys/net.py
@@ -1,23 +1,15 @@
-from litex.soc.integration.soc_core import mem_decoder
+from litex.soc.integration.soc_sdram import *
 
 from liteeth.core.mac import LiteEthMAC
 from liteeth.phy import LiteEthPHY
 
-from targets.utils import csr_map_update
-from targets.atlys.base import SoC as BaseSoC
+from .base import BaseSoC
 
 
 class NetSoC(BaseSoC):
-    csr_peripherals = (
-        "ethphy",
-        "ethmac",
-    )
-    csr_map_update(BaseSoC.csr_map, csr_peripherals)
-
-    mem_map = {
-        "ethmac": 0x30000000,  # (shadow @0xb0000000)
-    }
-    mem_map.update(BaseSoC.mem_map)
+    mem_map = {**BaseSoC.mem_map, **{
+        "ethmac": 0xb0000000,
+    }}
 
     def __init__(self, platform, *args, **kwargs):
         # Need a larger integrated ROM on or1k to fit the BIOS with TFTP support.
@@ -26,17 +18,25 @@ class NetSoC(BaseSoC):
 
         BaseSoC.__init__(self, platform, *args, **kwargs)
 
+        # Ethernet ---------------------------------------------------------------------------------
+        # Ethernet PHY
         self.submodules.ethphy = LiteEthPHY(
             platform.request("eth_clocks"),
-            platform.request("eth"),
-            self.clk_freq)
+            platform.request("eth"))
+        self.add_csr("ethphy")
 
+        # Ethernet MAC
+        ethmac_win_size = 0x2000
         self.submodules.ethmac = LiteEthMAC(
-            phy=self.ethphy, dw=32, interface="wishbone", endianness=self.cpu.endianness)
-        self.add_wb_slave(mem_decoder(self.mem_map["ethmac"]), self.ethmac.bus)
-        self.add_memory_region("ethmac",
-            self.mem_map["ethmac"] | self.shadow_base, 0x2000)
-
+            phy        = self.ethphy,
+            dw         = 32,
+            interface  = "wishbone",
+            endianness = self.cpu.endianness)
+        self.add_wb_slave(self.mem_map["ethmac"], self.ethmac.bus, ethmac_win_size)
+        self.add_memory_region("ethmac", self.mem_map["ethmac"], ethmac_win_size, type="io")
+        self.add_csr("ethmac")
+        self.add_interrupt("ethmac")
+        # timing constraints
         self.ethphy.crg.cd_eth_rx.clk.attr.add("keep")
         self.ethphy.crg.cd_eth_tx.clk.attr.add("keep")
         # FIXME: This is probably too tight?
@@ -44,7 +44,8 @@ class NetSoC(BaseSoC):
 
         self.platform.add_false_path_constraints(
             self.crg.cd_sys.clk,
-            self.ethphy.crg.cd_eth_rx.clk)
+            self.ethphy.crg.cd_eth_rx.clk,
+            self.ethphy.crg.cd_eth_tx.clk)
 
         self.platform.add_platform_command("""
 # FIXME: ERROR:Place:1108 - A clock IOB / BUFGMUX clock component pair have
@@ -57,8 +58,6 @@ NET "{eth_clocks_tx}" CLOCK_DEDICATED_ROUTE = FALSE;
             eth_clocks_rx=platform.lookup_request("eth_clocks").rx,
             eth_clocks_tx=platform.lookup_request("eth_clocks").tx,
             )
-
-        self.add_interrupt("ethmac")
 
     def configure_iprange(self, iprange):
         iprange = [int(x) for x in iprange.split(".")]
